Split index.tsx into smaller files
